Changed over to a 64-bit version in preparation to receive official art for the game. Currently using more test art scaled up to the 64x64 format.
Added zoom in and zoom out buttons to the HUD.

Removed the state switching method I implemented last commit. It's too buggy and works around the code too much. I need to find something that flows better.

Added a text file called "IN_PROGESS" to the src folder. This outlines some of the goals I aim to achieve with the TopDownState before moving on to implementing another state. It also outlines some of the things I'm currently working on with the code and where I'm stuck. It's something I plan on updating with each commit.