Cleaned up some of the action and zone switching related parts of code and did a bit of refactoring in that area. Still not completely happy with the design of using Strings and TextAction objects, but it does work a lot better than what was going on before.
Fixed the issue with the HUD rendering incorrectly when switching zones. I just had to handle the zone switching at the END of the rendering of everything else.

Currently implementing text actions. I have the HUD disappearing, text panel popping up, and player movement stopping. I still need to draw the text, allow for the player to click through the text (and possibly exit text), and bring back the original HUD when the TextAction is over. Needless to say, there is still a lot of work to be done with actions and refactoring the code. The action button is not implemented in this scenario as of yet.