Got Taylors practice map working on android devices.
This build has the current map working with the practice map, text actions, and changing states on desktop and android.

Reworked music so that it's handled more in the topDownState and is disposed of properly.

Added support so that FPS can be tracked on the screen; however, this is currently not working due to the nature of the FPSLogger class.